ActionHandlers now have a contract; refactored QueuePoller
